Modify delay function to utilize System Timer
I modified the code in the delay function to create a 100ms delay
utilizing the System Timer (SysTick). I also changed some variable and
subroutine names so they were all consistent with snake case naming
convention. I updated the README and added a picture of the physical
circuit as well.